requirements.txt
anyio==4.9.0
arrow==1.3.0
binaryornot==0.4.4
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cookiecutter==2.6.0
flet==0.27.6
flet-cli==0.27.6
flet-desktop==0.27.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
packaging==24.2
Pygments==2.19.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
rich==14.0.0
six==1.17.0
sniffio==1.3.1
text-unidecode==1.3
toml==0.10.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.1
urllib3==2.3.0
watchdog==4.0.2
google-generativeai>=0.3.0
pypdf>=3.15.1
langchain>=0.0.267
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
streamlit>=1.25.0
google-genai>=0.2.0
pyaudio>=0.2.13